// src/api/requests/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: string;
};

export type SignupRequestDto = {
    name: string;
    username: string;
    password: string;
    passwordConfirmation: string;
};

export type SignupResponseDto = {
    accessToken: string;
};

export type LoginRequestDto = {
    username: string;
    password: string;
};

export type LoginResponseDto = {
    accessToken: string;
};

export type SessionResponseDto = {
    id: string;
    username: string;
};

export type SignupData = {
    body: SignupRequestDto;
    path?: never;
    query?: never;
    url: '/auth/signup';
};

export type SignupResponses = {
    200: SignupResponseDto;
};

export type SignupResponse = SignupResponses[keyof SignupResponses];

export type LoginData = {
    body: LoginRequestDto;
    path?: never;
    query?: never;
    url: '/auth/login';
};

export type LoginResponses = {
    200: LoginResponseDto;
};

export type LoginResponse = LoginResponses[keyof LoginResponses];

export type GetSessionData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/auth/session';
};

export type GetSessionResponses = {
    200: SessionResponseDto;
};

export type GetSessionResponse = GetSessionResponses[keyof GetSessionResponses];

export type GetDebugInfoData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/debug';
};

export type GetDebugInfoResponses = {
    200: unknown;
};